Consolidate layouts
