+++ src/app/layout.tsx
@@ -1,0 +1,22 @@
+import './globals.css'
+
+export const metadata = {
+  manifest: '/manifest.json',
+  themeColor: '#000000',
+  title: 'Index App | Buy & Sell Our Tokens',
+  description:
+    'Use the Index Coop Trading App to buy and sell our sector, leveraged and yield generating tokens.',
+  type: 'website',
+}
+
+type LayoutProps = {
+  children: React.ReactNode
+}
+
+export default function RootLayout({ children }: LayoutProps) {
+  return (
+    <html lang='en'>
+      <body>{children}</body>
+    </html>
+  )
+}
